// src/utils/Enums.ts
export enum Gender{
  MALE = 'male',
  FEMALE = 'female',
  OTHERS = 'others'
}

export enum OrderStatus {
  CONFIRED = "CONFIRMED",      
  SHIPPED = "SHIPPED",         
  OUT_FOR_DELIVERY = "OUT_FOR_DELIVERY",
  DELIVERED = "DELIVERED",     
  CANCELLED = "CANCELLED",      
  RETURNED = "RETURNED",        
  REFUNDED = "REFUNDED"     
}